added literally everything
the website looks good now! rather than being a blank white page that says "hey :3" it's now a full, working linktree-type page.

it's based off my existing carrd, but this is intended to replace that. I'm not gonna delete the carrd just in case, but it won't be linked to in any of my major socials lol